ovs: fail port enslavement when the bridge AC is not found
The previous code tried to get the the bridge active connection and in
case of failure, it used the port active connection instead. This
doesn't seem right, as in nm-ovsdb.c the bridge AC is used to get the
bridge settings (including the uuid, interface name, and cloned mac).

In case of failure getting the bridge AC we should just fail.

Fixes: 830a5a14cb29 ('device: add support for OpenVSwitch devices')